+++ Changelog.md
@@ -9,6 +9,8 @@
 - Changed how the shop works
 - Made inventory easier to use
 - Fixed skill cooldowns not displaying properly
+- Changes to damage calculation
+- Fixed minor issues and improved QOL
 # 0.2a
 - Will add bank that can store up to 2/3 of exp. All exp not in bank is lost when defeated.
 - Added dodging to combat. Base chance is 5%, can eventually be increased with perks. Dodge nullifies damage and prevents statuses.